Rename directory service and config and added a unset token csrf
