+++ src/Service/CheckForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\service;
+namespace App\Service;
 
 class CheckForm
 {
